requirements.txt
nb-cli==0.6.7
aiohttp==3.8.3
cacheout==0.14.1
feedparser==6.0.10
loguru==0.6.0
nonebot2==2.0.0rc1
nonebot-adapter-onebot==2.1.3
nonebot-plugin-apscheduler==0.2.0
nonebot_plugin_apscheduler==0.2.0
numpy==1.23.5
opencv-python==4.6.0.66
opencv_python==4.6.0.66
playwright==1.27.1
psutil==5.9.4
pydantic==1.9.2
requests==2.28.1
oauthlib==3.2.2
async-lru==1.0.3
requests-oauthlib==1.3.1
tortoise-orm==0.19.2
tweepy==4.12.1
typing_extensions==4.4.0
yarl==1.8.1
bilibili-api-python==15.3.1
